use numel() instead of view and size. move skipIfNoLapack and TEST_SCIPY to test/common
